code update

+++ walker.py
@@ -142,8 +142,8 @@ class Walker:
                     arduino.write(command.encode('utf-8'))
                     time.sleep(0.1)
                 else:
-                    command = 'S1 ' + str(min(int(3000 * self.forback * self.left_turn_scale), 300) * (1 - 3 * self.leftright)) + \
-                              ',S2 ' + str(min(int(3000 * self.forback * self.right_turn_scale), 300)  * (1 - 3 * self.leftright)) + ",D1 0,D2 0,"
+                    command = 'S1 ' + str(int(min(int(3000 * self.forback * self.left_turn_scale), 300) * (1 - 3 * self.leftright))) + \
+                              ',S2 ' + str(int(min(int(3000 * self.forback * self.right_turn_scale), 300) * (1 - 3 * self.leftright))) + ",D1 0,D2 0,"
                     # arduino.write('S1 0,S2 0,D1 0,D2 0'.encode('utf-8'))
                     if command != self.precommand:
                         arduino.write(command.encode('utf-8'))
